Renamed 'project' folder to 'core'
