+++ nw/core/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+
+from nw.core.document import NWDoc
+from nw.core.index import NWIndex
+from nw.core.project import NWProject
+
+__all__ = [
+    "NWDoc",
+    "NWIndex",
+    "NWProject",
+]
