+++ nw/core/document.py
