--- nw/project/__init__.py
@@ -1,11 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from nw.project.document import NWDoc
-from nw.project.index import NWIndex
-from nw.project.project import NWProject
-
-__all__ = [
-    "NWDoc",
-    "NWIndex",
-    "NWProject",
-]
